+++ bin/_template/sample-post.js
@@ -1,0 +1,18 @@
+module.exports = {
+    "title": "Markdown Blog",
+    "author": "Your Name",
+    "date": new Date(),
+    "categories": [
+    ],
+    "tags": [
+    ],
+    "acceptComment": true,
+    "acceptTrackback": true,
+    "published": new Date(),
+    "status": "draft",
+    "important": false,
+    "advanced": {
+    }
+}
+
+