+++ config.js
@@ -1,56 +1,79 @@
-var conf = {
-    meta: {
-        version: '0.2.1',
-        defaultTitle: 'Markdown Blog - HarooGul',
-        description: 'Markdown Blogging engine based Node.js',
-        siteUrl: 'http://haroog.dev:8000',
-        author: 'Rhio Kim',
-		keywords: [ 'node.js', 'javascript', 'html5' ]
+module.exports = {
+    "meta": {
+        "version": "0.8.1",
+        "defaultTitle": "하루프레스 > Haroopress Official Site",
+        "description": "하루프레스는 하루를 손쉽게 기록해 보자라는 뜻의 블로그 엔진으로, Node.js 와 EJS 템플릿 엔진으로 설계된 Github 에서 제공        하는 정적 페이지 서비스를 위한 블로그 엔진입니다.",
+        "siteUrl": "http://haroopress.github.com",
+        "author": "Rhio Kim",
+        "keywords": [
+            "node.js",
+            "haroopress",
+            "octopress",
+            "wordpress",
+            "git",
+            "javascript",
+            "css",
+            "html",
+            "blog"
+        ]
     },
-    articles: process.cwd() +'/source/articles',
-    authors: process.cwd() +'/source/authors',
-
-    rss: {
-        contentLength: 0
+    "lang": "en",
+    "contentLength": 6,
+    "deployBranch": "gh-pages",
+    "CNAME": "",
+    "sourceDir": "/Users/rhio/Works/my/haroopress/source/data",
+    "publicDir": "/Users/rhio/Works/my/haroopress/_public",
+    "deployDir": "/Users/rhio/Works/my/haroopress/_deploy",
+    "themeDir": "/Users/rhio/Works/my/haroopress/source/themes",
+    "theme": {
+        "name": "basic"
     },
-    
-    sourceDir: __dirname + '/source',
-    deployDir: __dirname + '/_deploy',
-
-    deployBranch: 'gh-pages',
-    CNAME: 'nodejs.kr'
-};
-
-
-/* third party settings */
-conf.plugins = {
-    github: {
-        user: "rhiokim",
-        repoCount: 0,
-        profileLink: true,
-        skipForks: true
-    },
-    twitter: {
-        user: "rhiokim",
-        tweetCount: 4,
-        showReplies: false,
-        followerButton: true,
-        showFollowerCount: false,
-        tweetButton: false
-    },
-    google: {
-	googleAnalyticsId: 'UA-30492002-1',  //google analytics tracking id, default false
-	googlePlus: false,
-	googlePlusSize: 'medium'
-    },
-    disqus: {
-        shortName: "nodekr",
-        showCommentCount: false
-    },
-    facebook: {
-        showLikeButton: true 
-    },
-    delicious: {}
-};
-
-module.exports = conf;
+    "plugins": {
+        "github": {
+            "display": false,
+            "user": "",
+            "repoCount": 10,
+            "profileLink": true,
+            "skipForks": true
+        },
+        "twitter": {
+            "display": false,
+            "user": "",
+            "tweetCount": 10,
+            "showReplies": false,
+            "followerButton": false,
+            "showFollowerCount": false,
+            "tweetButton": false
+        },
+        "google": {
+            "display": false,
+            "googleAnalyticsId": "",
+            "googlePlus": false,
+            "googlePlusSize": "medium",
+            "gplus": {
+                "display": false,
+                "user": ""
+            }
+        },
+        "disqus": {
+            "display": false,
+            "shortName": "",
+            "showCommentCount": true
+        },
+        "facebook": {
+            "display": false,
+            "user": "",
+            "showSiteLink": false,
+            "showLikeButton": false
+        },
+        "delicious": {},
+        "contributors": {
+            "display": true
+        },
+        "weather": {
+            "display": false,
+            "delay": 0,
+            "zipcode": "KSXX0037"
+        }
+    }
+}