+++ History.md
@@ -1,4 +1,216 @@
 
+n.n.n / 2012-04-22 
+==================
+
+  * modify _config.js
+  * fixed css
+  * update rss template
+  * update wood theme css
+  * #105 google search plugin
+  * adjust theme to page
+  * #104 made wood-theme
+  * worked theme structure
+  * fixed public dir
+  * fixed public dir
+  * fixed public dir
+  * Merge branch 'feature/redesign-archive-list' into devel
+  * #96 redesign archives page
+  * object sorting test
+  * fixed conflict
+  * #93 disqus coment count is completed
+  * #97 is completed
+  * Merge branch 'devel' into feature/redesign-archive-list
+  * update js
+  * generated_toc remove
+  * add bg classname in well css
+  * worked basic markup & completed function
+  * worked twitter button link
+  * .
+  * enhancement header area
+  * Merge branch 'feature/404-page' into devel
+  * worked Issue #17
+  * merge from devel
+  * issue drop! gone to backlog
+  * Merge branch 'feature/plugin-github' into devel
+  * adjust mustache & worked github plugin, progress bar effect
+  * refactoring
+  * worked display repo detail info
+  * unregister repo click handler
+  * worked github & twitter base markup, function
+  * remove unused module
+  * worked make preview open browser
+  * fixed conflict
+  * Merge branch 'master' of github.com:rhiokim/haroopress into feature/404-page
+  * fixed confilict
+  * add to make init script mkdir public folder
+  * main.css -> haroo.css
+  * worked prev-next link
+  * working github repo plugin
+  * worked archive list
+  * Merge branch 'feature/archive-list' into devel
+  * worked all archives
+  * worked base markup
+  * remove public files & add gitnore pattern
+  * build step gen -> cp to public -> deploy from _deploy dir
+  * worked yahoo weather plugin
+  * bump
+  * worked 404 page markup
+  * worked new-page shell script
+  * update tab -> space
+  * update moment module & fixed date notation
+  * Merge branch 'feature/plugin-contributors' into devel
+  * update author articles information
+  * support contributors plugin
+  * .
+  * update changelog
+  * fixed page title at app.js
+  * add social button on content
+  * add social button on content
+  * update rss, twitter link
+  * page title, add author social info
+  * footer plugin
+  * add social icon, update config
+  * .
+  * remove garbege ejs views
+  * remove garbege cli
+  * update make script
+  * change repository name haroog -> haroopress
+  * Merge branch 'release/v0.7.0'
+  * Merge branch 'feature/page-generator' into devel
+  * worked status func for page
+  * worked sub navigation
+  * sub navgation structure test
+  * remove garbege image & update make gen & page structure
+  * page bump
+  * worked page generator & sample page
+  * bump
+  * shell script
+  * add findit module && create object model for shell script
+  * add findit module && create object model for shell script
+  * add stringEx module, add social icons, tutorial for haroopress
+  * remove old rss.xml & add favicon
+  * fixed #65 article link
+  * worked #35 support recent article plugin
+  * clear project & worked favorit-plugin
+  * Merge branch 'feature/author-list-page' into devel
+  * completed feature
+  * create author page & rss generating script
+  * worked author page markup & express app
+  * Merge branch 'feature/refactoring-cli' into devel
+  * app.js refactoring
+  * make gen2 test - success
+  * refactoring data control
+  * Merge branch 'feature/new-post-and-page' into devel
+  * delete template file
+  * #38 article generate cli script
+  * write new-post shell script
+  * del sample articles
+  * site info update
+  * move markdown template file & update footer
+  * dependency npm module add to make script
+  * update change log
+  * Merge branch 'release/v0.6.0'
+  * fixed git commit error
+  * delete comment
+  * twitter plugin & google plus plugin
+  * complete google analytics plugin
+  * Merge branch 'release/v0.5.1'
+  * fixed category at article page
+  * deploy test, disqus plugin activated
+  * add deploy make script
+
+n.n.n / 2012-04-12 
+==================
+
+  * worked #35 support recent article plugin
+  * clear project & worked favorit-plugin
+  * Merge branch 'feature/author-list-page' into devel
+  * completed feature
+  * create author page & rss generating script
+  * worked author page markup & express app
+  * Merge branch 'feature/refactoring-cli' into devel
+  * app.js refactoring
+  * make gen2 test - success
+  * refactoring data control
+  * Merge branch 'feature/new-post-and-page' into devel
+  * delete template file
+  * #38 article generate cli script
+  * write new-post shell script
+  * del sample articles
+  * site info update
+  * move markdown template file & update footer
+  * dependency npm module add to make script
+  * update change log
+  * fixed git commit error
+  * delete comment
+  * twitter plugin & google plus plugin
+  * complete google analytics plugin
+  * fixed category at article page
+  * deploy test, disqus plugin activated
+  * add deploy make script
+  * clear devel
+  * submodule update && gh-pages-deployer completed
+  * push
+  * create sub category pages generator
+  * setup github gh-pages
+  * complete sub category generator & category menu error fixed
+  * Merge branch 'master' of github.com:rhiokim/haroog into devel
+  * Merge branch 'devel' of github.com:rhiokim/haroog into devel
+  * update layout, cate page
+  * Merge branch 'master' of https://github.com/rhiokim/haroog
+  * Merge branch 'release/v0.4.0'
+  * Merge branch 'feature/static-generator' into devel
+  * complete static generator
+  * adjust bootstrap v2.0.2
+  * index.html generator & make gen script update
+  * Merge branch 'devel' of github.com:rhiokim/haroog into devel
+  * Merge branch 'devel' of https://github.com/rhiokim/haroog into feature/static-generator
+  * update readme
+  * add directory generator
+  * add mkdirp module
+  * add ejs module
+  * create changelog
+  * Merge branch 'release/v0.3.1'
+  * update markdown-blog module
+  * Merge branch 'release/v0.3.0'
+  * fixed undefined error
+  * responsive header!
+  * gitignore atlassian~
+  * clear repo submodule error & update twitter bootstrap v2.0.2
+  * unregister direc & markdown-blog submodule
+  * update build script for generator & twitter bootstrap auto update
+  * getUserMedia idea 정리
+  * create update & fixed node.js category link
+  * Merge branch 'feature/categories-page' into devel
+  * complete category page
+  * written category markup
+  * written category generator cli
+  * add category.ejs
+  * fixed markup miss
+  * Merge branch 'devel' of https://github.com/rhiokim/haroog into devel
+  * remove direc module in app.js
+  * Merge branch 'devel' of https://github.com/rhiokim/haroog
+  * update readme
+  * Merge branch 'release/v0.2.0'
+  * Merge branch 'feature/support-rss' into devel
+  * complete rss.xml generator
+  * Merge branch 'release/v0.1.0'
+  * Merge branch 'feature/plugins' into devel
+  * modify nanhapark project
+  * complete plugin structure
+  * merge from devel
+  * add gen, preview make script
+  * layout design update
+  * Merge branch 'devel' into feature/plugins
+  * update readme.markdown
+  * update readme.markdown
+  * update reamde.markdown
+  * add intro png
+  * update readme.markdown
+  * update reamde.markdown
+  * 레이아웃 디자인 변경, 소셜 버튼 추가
+  * 레이아웃 디자인 변경, 소셜 버튼 추가
+
 0.6.0 / 2012-03-31 
 ==================
 
@@ -35,7 +247,6 @@
   * update readme
   * Merge branch 'release/v0.2.0'
   * Merge branch 'feature/support-rss' into devel
-  * .
   * complete rss.xml generator
   * Merge branch 'release/v0.1.0'
   * Merge branch 'feature/plugins' into devel